wip: fixes #14: parser for kubectl command line
Parser holds the syntax tree of the command line. The class is used to
read the json schema into a tree structure. The class is also used to
identify the tokens in a command line by walking the tree and returning
a sequence of possible suggestions at any given node in the walk

- print_tree(..) can be used to print the command tree
- autocomplete local options at every subcmd
- completeing globalFlags at each level
- adds debug logging that can help troubleshoot treewalks
- handling the case when --opt=value type input is given
- handling values for options and arguments and maintaining autocomlpete