Add flake8 to the testing, Python 2.7.12 -> 2.7.13
I do not know if this of interest to you but the goal of this change is to have Travis-CI automatically run flake8 tests on every pull request. This will help contributors know if their submissions are going to break the build. To turn on this free service, the repo owner would need to do steps 1 and 2 of https://docs.travis-ci.com/user/getting-started/